settings_engine/requirements.txt
greenlet==1.1.2
numpy==1.21.5
pandas==1.3.5
psycopg2-binary==2.9.2
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.29
python-dotenv==0.19.2
